requirements.txt
requests
matplotlib
python-dateutil
pandas
numpy
tensorflow 
flask
flask_cors
dash
dash-bootstrap-components
plotly
fastapi 
uvicorn 
jinja2